+++ nachos/qcip_tools_ext/gaussian.py
@@ -1,0 +1,18 @@
+from qcip_tools.chemistry_files import gaussian
+from qcip_tools.chemistry_files import PropertyNotPresent
+
+
+@gaussian.FCHK.define_property('n:input_electric_field')
+def gaussian__FCHK__get_input_electric_field(obj, *args, **kwargs):
+    """Get the input electric field
+
+    :param obj: object
+    :type obj: qcip_tools.chemistry_files.gaussian.FCHK
+    :rtype: dict
+    """
+
+    if 'External E-field' in obj:
+        field = obj.get('External E-field')
+        return field
+    else:
+        raise PropertyNotPresent('n:input_electric_field')
